feat/chatStyle : 기본 레이아웃
