+++ client/package.json
@@ -9,6 +9,8 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.52.1",
+    "sass-loader": "^13.0.0",
     "socket.io-client": "^4.5.1",
     "web-vitals": "^2.1.4"
   },
