+++ client/src/components/chat/chat.scss
@@ -1,0 +1,108 @@
+// npm i sass 필요
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap');
+
+// Container
+.chatBlockStyle {
+    display : flex;
+    flex-direction: column;
+    justify-content: space-between;
+    position : absolute;
+    width : 350px;
+    height : 100vh;
+    left : calc(100vw - 350px);
+    top : 0px;
+    background-color : #3a1c54;
+
+    form {
+        width : 85%;
+        margin : 0 auto;
+        display : flex;
+        justify-content: space-between;
+        padding : 2rem 0;
+
+        input {
+            border-radius: 5px;
+            font-size : 1rem;
+            width : 200px;
+            padding : 1.5rem 1rem;
+        }
+    
+        button {
+            border-radius: 5px;
+            padding : 1rem;
+        }
+    }
+}
+
+//채팅 log Container
+.chatContainer {
+    margin-top : 1rem;
+    height : 100%;
+    border-radius: 5px;
+    
+}
+
+// 채팅 log
+.chatStyle {
+    width : 85%;
+    height : 100%;
+    margin : 0 auto;
+    background-color : #250f39;
+    border-radius: 5px;
+    overflow-x : hidden;
+    overflow-y : scroll;
+
+    // 스크롤바
+    &::-webkit-scrollbar {
+        width : 10px;
+        
+    }
+
+    &::-webkit-scrollbar-thumb {
+        background-color : #f2c76f;
+        border-radius: 10px;
+        background-clip: padding-box;
+        border: 2px solid transparent;
+    }
+
+    &::-webkit-scrollbar-track {
+        display : none;
+    }
+    
+    // 글자 기본값
+    p {
+        &:first-child {
+            padding-top : 1rem;
+        }
+
+        padding : 0;
+        font-size : 1rem;
+        font-weight :normal;
+        color : #eee;
+    }
+
+    // 닉네임
+    .Nickname {
+        font-size : 1.1rem;
+        font-weight : 500;
+        padding : .5rem 1rem .2rem;
+    }
+
+    // 채팅 내용
+    .Content {
+        font-weight : 300;
+        padding-top : 0;
+        padding : 0 1rem .5rem;
+    }
+}
+
+// 입력 log
+.formStyle {
+    background-color : yellow;
+    position: "relative";
+    top : calc(100vh - 30px);
+    left : calc(100vw - 1450px);
+    height : 100%;
+    min-height : 200px;
+}
+
