+++ client/src/index.css
@@ -1,3 +1,11 @@
+/* 기본값 세팅 */
+* {
+  margin : 0;
+  padding : 0;
+  overflow : hidden;
+  font-family: 'Noto Sans KR', sans-serif;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
